requirements.txt
datasets==2.4.0
numpy==1.19.5
transformers==4.18.0
torch==1.10.2
scikit-learn==0.24.2
fastapi==0.79.0
pydantic==1.9.1
uvicorn==0.16.0
PyYAML==6.0
